change file and fixing on the file
