+++ front-end/src/components/layout/NavbarContainer.js
@@ -1,7 +1,7 @@
 // import axios from 'axios';
 import React from 'react';
-import NavbarisLogin from './NavbarisLogin';
-import NavbarNotLogin from './NavbarNotLogin';
+import NavbarisLogin from './navbarisLogin';
+import NavbarNotLogin from './navbarNotLogin';
 // import { navDes } from '../pages/style';
 // import { useNavigate } from 'react-router-dom';
 // import { useMatch } from 'react-router-dom';
