+++ front-end/src/components/layout/mainLayout.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Navbar from './NavbarContainer';
+import Navbar from './navbarContainer';
 import Footer from './footer';
 const MainLayouts = (props) => {
   const { children } = props;
